[cleanup] [minor] added webnotes.model.delete_doc (module)
